Allow const Constraints in constructors
Constraints are only used in const-safe ways, so allow a caller
to pass a const Constraint.

This means firstly it's clearer that a constraint won't be modified
by use by an Arg object, so you can safely share a constraint between
different args.

It also allows static construction of constant constraints.