Fix installation of doc files.
Generation of doc files depends on a CMake option BUILD_DOC and if CMake can find Doxygen and/or LibXslt.
The installation of the doc files however is undonditionally.
With this patch doc files are only installed if peviously generated.

Patch submitted by frogarian@ in https://sourceforge.net/p/tclap/patches/19/